backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
python-dotenv==1.0.1
djangorestframework-simplejwt==4.7.2